refactor(agent): dissolve messages/helpers runtime clusters into their single callers
The replay-merge cluster (mergeCanonicalAndRuntimeEvents,
reconcileReplayMessages and their private helpers) was imported only by
runtime/engine.ts, and the queue-reconciliation cluster
(reconcileQueueWithPiEvent, removeDeliveredQueuedMessage and theirs) only
by runtime/pi-event-applier.ts. Move each cluster into its sole caller as
private functions so runtime-lifecycle reading no longer bounces through
messages/helpers.ts. Inline runtimeStatusLooksActive at its one
prompt-stream call site and delete the identity visibleQueuedMessages,
keeping its WHY comment at the chat-pane call site. queueDisplayText keeps
importing the shared visibleUserTextFromPi.
