+++ frontend/src/features/agent/messages/helpers.ts
@@ -7,12 +7,7 @@ import {
 } from "@shared/agent/session-title";
 
 export { cleanSessionTitle, isPlaceholderSessionTitle };
-import type {
-  ChatMessage,
-  QueuedMessage,
-  RuntimeLoggedEvent,
-  TokenStats,
-} from "@/features/agent/messages/types";
+import type { TokenStats } from "@/features/agent/messages/types";
 import type { Session } from "@/features/agent/runtime/types";
 
 export function randomIdSegment(length: number): string {
@@ -157,10 +152,6 @@ export function messageText(
     .join(separator);
 }
 
-export function runtimeStatusLooksActive(status: { active?: boolean }): boolean {
-  return status.active === true;
-}
-
 /** Whether a runtime status snapshot says the session can take a steer or a
  *  follow-up right now.
  *
@@ -197,176 +188,6 @@ export function replayCursorAfterRuntimeHydration(
   const matchesSession = runtimeStatus.piSessionId === piSessionId;
   const activeUnclaimed = runtimeStatus.active === true && !runtimeStatus.piSessionId;
   return matchesSession || activeUnclaimed ? runtimeStatus.eventSeq : undefined;
-}
-
-/** Every item still in the queue is pending delivery, so all of them show.
- *
- * This used to hide `sent` items, which meant EVERY follow-up — they are
- * marked sent the moment pi accepts them — so the drawer stack was always
- * empty and queueing looked broken. Items leave the queue when pi actually
- * delivers them (the user echo) or contradicts us (`queue_update`). */
-export function visibleQueuedMessages(queue: QueuedMessage[]): QueuedMessage[] {
-  return queue;
-}
-
-function stringArray(value: unknown): string[] {
-  return Array.isArray(value)
-    ? value.filter((item): item is string => typeof item === "string" && item.trim().length > 0)
-    : [];
-}
-
-function queueDisplayText(text: string): string {
-  return visibleUserTextFromPi(text) || text.trim();
-}
-
-function queueKey(mode: QueuedMessage["mode"], text: string): string {
-  return `${mode}:${queueDisplayText(text)}`;
-}
-
-function consumePending(
-  pending: Map<string, string[]>,
-  mode: QueuedMessage["mode"],
-  text: string,
-): string | null {
-  const key = queueKey(mode, text);
-  const values = pending.get(key);
-  if (!values || values.length === 0) return null;
-  const [value, ...remaining] = values;
-  if (remaining.length > 0) pending.set(key, remaining);
-  else pending.delete(key);
-  return value ?? null;
-}
-
-export function reconcileQueueWithPiEvent(
-  queue: QueuedMessage[],
-  event: Record<string, unknown>,
-): QueuedMessage[] {
-  if (event.type !== "queue_update") return queue;
-  const pending = new Map<string, string[]>();
-  const addPending = (mode: QueuedMessage["mode"], messages: string[]) => {
-    for (const text of messages) {
-      const key = queueKey(mode, text);
-      pending.set(key, [...(pending.get(key) ?? []), text]);
-    }
-  };
-  addPending("follow_up", stringArray(event.followUp));
-
-  const next = queue.flatMap((item) => {
-    if (item.mode !== "follow_up") return [];
-    const acceptedByPi = consumePending(pending, item.mode, item.text);
-    if (acceptedByPi) return [{ ...item, text: queueDisplayText(acceptedByPi), sent: true }];
-    return item.sent ? [] : [item];
-  });
-
-  for (const [key, messages] of pending) {
-    const separator = key.indexOf(":");
-    const mode = key.slice(0, separator) as QueuedMessage["mode"];
-    for (const text of messages) {
-      next.push({ id: newId("queue"), mode, text: queueDisplayText(text), sent: true });
-    }
-  }
-  return next;
-}
-
-export function removeDeliveredQueuedMessage(
-  queue: QueuedMessage[],
-  deliveredText: string,
-): QueuedMessage[] {
-  const delivered = queueDisplayText(deliveredText);
-  const index = queue.findIndex((item) => queueDisplayText(item.text) === delivered);
-  if (index === -1) return queue;
-  return [...queue.slice(0, index), ...queue.slice(index + 1)];
-}
-
-function eventKey(event: Record<string, unknown>): string {
-  try {
-    return JSON.stringify(event);
-  } catch {
-    return `${String(event.type ?? "event")}:${Object.keys(event).join(",")}`;
-  }
-}
-
-function messageFingerprint(event: Record<string, unknown>): string | null {
-  const message = asRecord(event.message);
-  if (!message || typeof message.role !== "string") return null;
-  return eventKey(message);
-}
-
-function canonicalEventsBeforeRuntimeTail(
-  canonicalEvents: Record<string, unknown>[],
-  runtime: Record<string, unknown>[],
-): Record<string, unknown>[] {
-  const canonicalMessages = canonicalEvents.flatMap((event, eventIndex) => {
-    const fingerprint = messageFingerprint(event);
-    return fingerprint ? [{ eventIndex, fingerprint }] : [];
-  });
-  const runtimeMessages = runtime.flatMap((event) => {
-    if (event.type !== "message" && event.type !== "message_end") return [];
-    const fingerprint = messageFingerprint(event);
-    return fingerprint ? [fingerprint] : [];
-  });
-  const firstRuntimeMessage = runtimeMessages[0];
-  if (!firstRuntimeMessage) return canonicalEvents;
-  let best: { eventIndex: number; score: number } | null = null;
-  for (let index = 0; index < canonicalMessages.length; index += 1) {
-    if (canonicalMessages[index]?.fingerprint !== firstRuntimeMessage) continue;
-    let score = 0;
-    while (
-      canonicalMessages[index + score]?.fingerprint === runtimeMessages[score] &&
-      runtimeMessages[score]
-    ) {
-      score += 1;
-    }
-    const candidate = { eventIndex: canonicalMessages[index]?.eventIndex ?? 0, score };
-    if (!best || candidate.score >= best.score) best = candidate;
-  }
-  if (best) {
-    return canonicalEvents.slice(0, best.eventIndex);
-  }
-  return canonicalEvents;
-}
-
-function runtimeEventsInOrder(
-  runtimeEvents: readonly RuntimeLoggedEvent[],
-): Record<string, unknown>[] {
-  return [...runtimeEvents]
-    .sort((a, b) => (a.seq ?? 0) - (b.seq ?? 0))
-    .flatMap((entry) => {
-      if (entry.event && typeof entry.event === "object") {
-        return [entry.event];
-      }
-      return [];
-    });
-}
-
-function dedupeAdjacentEvents(events: Record<string, unknown>[]): Record<string, unknown>[] {
-  let previous = "";
-  return events.filter((event) => {
-    const key = eventKey(event);
-    if (key === previous) return false;
-    previous = key;
-    return true;
-  });
-}
-
-export function mergeCanonicalAndRuntimeEvents(
-  canonicalEvents: Record<string, unknown>[],
-  runtimeEvents: readonly RuntimeLoggedEvent[] = [],
-): Record<string, unknown>[] {
-  const runtime = runtimeEventsInOrder(runtimeEvents);
-  return dedupeAdjacentEvents([
-    ...canonicalEventsBeforeRuntimeTail(canonicalEvents, runtime),
-    ...runtime,
-  ]);
-}
-
-export function reconcileReplayMessages(
-  current: ChatMessage[],
-  canonical: ChatMessage[],
-): ChatMessage[] {
-  if (canonical.length === 0) return current;
-  if (canonical.length >= current.length) return canonical;
-  return current;
 }
 
 export function makeFreshTab(): Session {
